Added the offer section component
This component takes no props. It focuses on gaining info from offers
collection and constructing offers section containing all the offers.
There are no styles provided as of yet but in the future a slider
function will be implemented for a offers carousel. This will be done in
the styles feature update in the later stage of development