Moved and modified OfferSection
The OfferSection component did not fit the description of a component
due to which it was shifted to Areas. This ensured the clarity of
components.

The Changes made to OfferSection was made to modify the was OfferSection
gains information. Before the getOffers method was directly inBuilt in
OfferSection But now the Offer section is provided the value in Context
values. This helps in keeping the offer object consistant across the app
and let other components use it.