Added contexts to index
The contexts Sections and Offers were added to index this was done to
ensure that the api call to firebase remain minimal and all the
components which use section or offer can recive the data at the same
time.
It will further reduce the load time of the page after the first load.