Added a Offers context to the app
This initially provides components with empty Object with two feilds
groups and banners (A seperation created to distinguish between a slider
offer and a mainline offer) after which it retrieves the data from the
firestore database and arranges the data w.r.t. two feilds and then
provides it to the components. This keeps the initial load time of the
app as low as possible while providing the apps with the required data
after the first repaint of the page.