Added a SectionContext
This provides components with the sections data stored on firestore and
ensures that the data is only recieved once in the app. It provides user
with a better experince by splitting the load time in 2 parts the the
initial loading time which focuses on loading the static assets and the
initial scripts and the later loading time in which it can focus totally
on providing information to the app.