Page Styles Update
This update includes the styling of various components of the home page
this includes:
	- providing a global style
	- providing individual component style
	- providing a single component with multiple styles with the
	  help of class names

GLOBAL STYLES:
This includes all the default styles for elements such as h1,h2,etc. It
also provides the user with style variables such as colors etc. These
variables are that widely used yet but will be in a future update.

OTHER STYLES:
These are component and area styles which are provided fot individual
sections of the page.

TAG:
The divs were converted to their respectful tags such as
header/footer/main etc.

Footer:
The
The next update will add functionality to the page such as how
components can be interacted with and it will use styles extensively.