Removed the offer styles and changed the slider
The Group offer now renders a fragment of array of StyledGroupOffer
elements. Now this Groupoffer element consists the title of the entire
section followed by an offer slider.

The OfferSlider in the meanwhile has been updated to provide only with
code related to an Offer slider it focuses on creating an interactable
slider component to display all the offers in the group as Offer
component. The styled is also now focused on providing a Offer slider
and all the redundant styling had been removed.