Added the controls and functionality for banner
This commit focuses on providing a relation between the buttons and the
image div translation. various methods such as setTimeout were
considered. This implementation keep the code as simple as possible.

The buttons are mapped in accordance to the no. of offers present. Each
Button has an animation running for a duration which is specified in
OfferBanner. When this animation ends an animationEnd event is triggered
which leads to the change of the current state.

The state itself focuses on how much translation to provide to the
Banner component for its proper functioning.

This banner can be used for dynamic number of images.