StyledOffer was replaced & button was implemented
The styled offer was converted to StyledSliderComponent due to the lack
of need for the said component. The new component is supposed to be a
child div of the StyledOffer and handles all the scroll related styles.

slider button was implemented to control the sliding effect of
StyledSliderComponent all the styles are provided in this file.