Added variables to handle the scroll
The various variables and functions handle the scroll component. The
implementation is done by using ref and events.

The isInactive state has 2 variables left and right which informs the
component about which button to deactivate depending on the scroll
position.

The assignment of this state is handled by scrollController which
focuses on handling the scroll event on the given divs.

The left and right functions are used to change the current scroll
position of the element. The intensity of this scroll can be adjusted by
ScrollLength variable set at the start of the function.